Move user's authentication schema to the UserSchema class
